feat(SMART-DATA-02): add new network configuration to Docker and implement PostgreSQL support
- Add 'smartdata-network' to docker-compose.
- Implement new PostgreSQL database service.
- Change reading value type from Integer to Double in reading classes.
- Implement synthetic data generation for humidity and temperature in the database.
- Implement generation and sending of temperature and humidity data in the Producer.

+++ producer/src/main/java/com/producer/dto/ReadingDTO.java
@@ -3,6 +3,6 @@ package com.producer.dto;
 import java.sql.Timestamp;
 
 public record ReadingDTO(
-    Integer value,
+    Double value,
     Timestamp timestamp){
 }
